test: add test for AddEmail with error
